requirements.txt
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
catboost==1.2.7
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flask==3.0.3
fonttools==4.54.1
graphviz==0.20.3
h11==0.14.0
humanize==4.10.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipyvue==1.11.1
ipyvuetify==1.10.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.60.0
numpy==1.26.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt-toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
reacton==1.8.3
referencing==0.35.1
requests==2.32.3
rich==13.9.1
rich-click==1.8.3
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.13.1
shap==0.46.0
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
solara==1.39.0
solara-server==1.39.0
solara-ui==1.39.0
stack-data==0.6.3
starlette==0.39.2
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
werkzeug==3.0.4
widgetsnbextension==4.0.13
zipp==3.20.2
